training/requirements.txt
ultralytics>=8.0.0
mlflow>=2.10.0
pyyaml>=6.0
onnx>=1.14.0
onnxruntime>=1.16.0
onnxsim>=0.4.33
